Feat: TicketItem 컴포넌트 생성 및 types 폴더 분리
