+++ src/types/schedule/schedule.ts
